Fix bug in client's KeyboardHandler in encodeCommand -> Opcode fromString instead of fromInt

+++ server/src/main/java/bgu/spl/net/impl/tftp/TftpEncoderDecoder.java
@@ -8,7 +8,7 @@ import bgu.spl.net.api.MessageEncoderDecoder;
 public class TftpEncoderDecoder implements MessageEncoderDecoder<byte[]> {
 
     private List<Byte> bytes = new ArrayList<Byte>();
-    private byte op_code = 0;
+    private short op_code = 0;
     private boolean messageEndsWithZero = true;
     private boolean messageEnded = false;
     private short dataBytesLeft;
